Fix dtype mismatch by converting tensors to float32
- Add .float() to np2tensor to ensure float32 dtype
- Prevents RuntimeError: mat1 and mat2 dtype mismatch
- NumPy defaults to float64, PyTorch networks use float32

+++ src/sim/train_agent.py
@@ -35,7 +35,7 @@ def _flatten_dict(d, parent_key="", sep="_"):
 
 
 def np2tensor(np_dict: dict) -> dict:
-    return {k: torch.from_numpy(v) for k, v in np_dict.items()}
+    return {k: torch.from_numpy(v).float() for k, v in np_dict.items()}
 
 
 def tensor2numpy(tensor_dict: dict) -> dict:
